+++ lib/basic_utils.dart
@@ -17,6 +17,7 @@ export 'src/model/pkcs7/Pkcs7CertificateData.dart';
 export 'src/model/x509/X509CertificateData.dart';
 export 'src/model/x509/X509CertificateObject.dart';
 export 'src/model/x509/X509CertificateValidity.dart';
+export 'src/model/x509/ExtendedKeyUsage.dart';
 export 'src/model/csr/CertificateSigningRequestData.dart';
 export 'src/model/DnsApiProvider.dart';